Implement new http base connector
