+++ mex/common/connector/base.py
